[Drop-In UI] Added ability to show Info Panel in Free Drive mode. (#6011)
* Introduced NavigationViewOptions.showInfoPanelInFreeDrive that allows InfoPanel to be shown in FreeDrive state.
Fixed bottom sheet initial, hidden state layout location.
QA Test App - Cleanup + Drawer layout

* CHANGELOG and metalava api file.
Removed debug view.

* KDoc updates.

* KDoc updates and QA App Drawer button icon tweaks.
