+++ libnavui-dropin/src/main/res/layout/mapbox_info_panel_header_layout.xml
@@ -30,6 +30,7 @@
             android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:layout_marginTop="20dp"
+            android:visibility="gone"
             app:layout_constraintTop_toTopOf="parent"
             app:layout_constraintEnd_toStartOf="@id/tripProgressGuideline"
             app:layout_constraintStart_toStartOf="parent" />
@@ -56,6 +57,8 @@
             android:layout_height="52dp"
             android:layout_marginTop="26dp"
             android:layout_marginEnd="10dp"
+            android:visibility="gone"
+            tools:visibility="visible"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintTop_toTopOf="parent" />
 
@@ -68,6 +71,8 @@
             android:text="POI Name"
             android:maxLines="2"
             android:ellipsize="end"
+            android:visibility="gone"
+            tools:visibility="visible"
             android:layout_marginStart="10dp"
             android:layout_marginEnd="10dp"
             app:layout_constraintStart_toStartOf="parent"
@@ -80,6 +85,8 @@
             android:layout_width="72dp"
             android:layout_height="52dp"
             android:layout_marginEnd="10dp"
+            android:visibility="gone"
+            tools:visibility="visible"
             app:layout_constraintEnd_toStartOf="@+id/startNavigation"
             app:layout_constraintBottom_toBottomOf="@+id/startNavigation"
             app:layout_constraintTop_toTopOf="@+id/startNavigation" />
@@ -90,6 +97,8 @@
             android:layout_height="52dp"
             android:layout_marginTop="26dp"
             android:layout_marginEnd="10dp"
+            android:visibility="gone"
+            tools:visibility="visible"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintTop_toTopOf="parent" />
