+++ qa-test-app/src/main/java/com/mapbox/navigation/qa_test_app/view/MapboxNavigationViewActivity.kt
@@ -1,16 +1,38 @@
 package com.mapbox.navigation.qa_test_app.view
 
 import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
+import android.view.View
 import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
 import com.mapbox.navigation.qa_test_app.databinding.LayoutActivityNavigationViewBinding
+import com.mapbox.navigation.qa_test_app.databinding.LayoutDrawerMenuNavViewBinding
+import com.mapbox.navigation.qa_test_app.view.base.DrawerActivity
 
-@ExperimentalPreviewMapboxNavigationAPI
-class MapboxNavigationViewActivity : AppCompatActivity() {
+@OptIn(ExperimentalPreviewMapboxNavigationAPI::class)
+class MapboxNavigationViewActivity : DrawerActivity() {
+
+    private lateinit var binding: LayoutActivityNavigationViewBinding
+    private lateinit var menuBinding: LayoutDrawerMenuNavViewBinding
+
+    override fun onCreateContentView(): View {
+        binding = LayoutActivityNavigationViewBinding.inflate(layoutInflater)
+        return binding.root
+    }
+
+    override fun onCreateMenuView(): View {
+        menuBinding = LayoutDrawerMenuNavViewBinding.inflate(layoutInflater)
+        return menuBinding.root
+    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        val binding = LayoutActivityNavigationViewBinding.inflate(layoutInflater)
-        setContentView(binding.root)
+
+        menuBinding.toggleReplay.isChecked = binding.navigationView.api.isReplayEnabled()
+        menuBinding.toggleReplay.setOnCheckedChangeListener { _, isChecked ->
+            if (isChecked) {
+                binding.navigationView.api.enableReplaySession()
+            } else {
+                binding.navigationView.api.enableTripSession()
+            }
+        }
     }
 }
